people-and-planet-ai/land-cover-classification/src/trainer-pytorch/pyproject.toml
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# https://peps.python.org/pep-0621
[project]
name = "trainer-pytorch"
version = "1.0.0"
dependencies = [
  "lightning==2.0.9",
  "tensorboardX==2.6.2.2",
  "torchvision==0.15.2",
]

[project.optional-dependencies]
vertex = ["google-cloud-aiplatform==1.33.1"]
